+++ cedar-policy/CHANGELOG.md
@@ -20,6 +20,9 @@ Starting with version 3.2.4, changes marked with a star (*) are _language breaki
 ### Added
 - Public syntax tree (`pst`) support for `variadic-is-in-range` feature: a variadic `isInRange` is modelled by a `pst::Expr::VariadicOp{...}` in the PST (#2380).
 - Functions for inspecting TPE policy evaluation results (`TpeResponse::reason` and iterators for true/false/error/residual permit/forbid policy IDs).
+- Depth limited parsing functions for `PolicySet`, `Template`, `Policy`, and `Expression`. These allow configuring a depth limit
+  for policies, avoiding stack overflows when parsing policies of unknown depth, assuming that the limit is set low enough to
+  avoid exhausting stack space in your environment. (#2383)
 
 ### Changed